split dumpling utils
