+++ dm/pkg/dumpling/utils_test.go
@@ -15,12 +15,15 @@ package dumpling
 
 import (
 	"os"
+	"strings"
 	"testing"
 
 	"github.com/go-mysql-org/go-mysql/mysql"
 	. "github.com/pingcap/check"
+	"github.com/pingcap/tidb/dumpling/export"
 
 	"github.com/pingcap/ticdc/dm/pkg/gtid"
+	"github.com/pingcap/ticdc/dm/pkg/log"
 	"github.com/pingcap/ticdc/dm/pkg/terror"
 )
 
@@ -254,4 +257,68 @@ Finished dump at: 2020-12-02 17:13:56
 	c.Assert(err, IsNil)
 	_, _, err = ParseMetaData(f.Name(), "mysql")
 	c.Assert(terror.ErrMetadataNoBinlogLoc.Equal(err), IsTrue)
+}
+
+func (t *testSuite) TestParseArgs(c *C) {
+	logger := log.L()
+
+	exportCfg := export.DefaultConfig()
+	extraArgs := `--statement-size=100 --where t>10 --threads 8 -F 50B`
+	err := ParseExtraArgs(&logger, exportCfg, strings.Fields(extraArgs))
+	c.Assert(err, IsNil)
+	c.Assert(exportCfg.StatementSize, Equals, uint64(100))
+	c.Assert(exportCfg.Where, Equals, "t>10")
+	c.Assert(exportCfg.Threads, Equals, 8)
+	c.Assert(exportCfg.FileSize, Equals, uint64(50))
+
+	extraArgs = `--threads 16 --skip-tz-utc`
+	err = ParseExtraArgs(&logger, exportCfg, strings.Fields(extraArgs))
+	c.Assert(err, NotNil)
+	c.Assert(exportCfg.Threads, Equals, 16)
+	c.Assert(exportCfg.StatementSize, Equals, uint64(100))
+
+	// no `--tables-list` or `--filter` specified, match anything
+	c.Assert(exportCfg.TableFilter.MatchTable("foo", "bar"), IsTrue)
+	c.Assert(exportCfg.TableFilter.MatchTable("bar", "foo"), IsTrue)
+
+	// specify `--tables-list`.
+	extraArgs = `--threads 16 --tables-list=foo.bar`
+	err = ParseExtraArgs(&logger, exportCfg, strings.Fields(extraArgs))
+	c.Assert(err, IsNil)
+	c.Assert(exportCfg.TableFilter.MatchTable("foo", "bar"), IsTrue)
+	c.Assert(exportCfg.TableFilter.MatchTable("bar", "foo"), IsFalse)
+
+	// specify `--tables-list` and `--filter`
+	extraArgs = `--threads 16 --tables-list=foo.bar --filter=*.foo`
+	err = ParseExtraArgs(&logger, exportCfg, strings.Fields(extraArgs))
+	c.Assert(err, ErrorMatches, ".*cannot parse --tables-list and --filter together.*")
+
+	// only specify `--filter`.
+	extraArgs = `--threads 16 --filter=*.foo`
+	err = ParseExtraArgs(&logger, exportCfg, strings.Fields(extraArgs))
+	c.Assert(err, IsNil)
+	c.Assert(exportCfg.TableFilter.MatchTable("foo", "bar"), IsFalse)
+	c.Assert(exportCfg.TableFilter.MatchTable("bar", "foo"), IsTrue)
+
+	// compatibility for `--no-locks`
+	extraArgs = `--no-locks`
+	err = ParseExtraArgs(&logger, exportCfg, strings.Fields(extraArgs))
+	c.Assert(err, IsNil)
+	c.Assert(exportCfg.Consistency, Equals, "none")
+
+	// compatibility for `--no-locks`
+	extraArgs = `--no-locks --consistency none`
+	err = ParseExtraArgs(&logger, exportCfg, strings.Fields(extraArgs))
+	c.Assert(err, IsNil)
+	c.Assert(exportCfg.Consistency, Equals, "none")
+
+	extraArgs = `--consistency lock`
+	err = ParseExtraArgs(&logger, exportCfg, strings.Fields(extraArgs))
+	c.Assert(err, IsNil)
+	c.Assert(exportCfg.Consistency, Equals, "lock")
+
+	// compatibility for `--no-locks`
+	extraArgs = `--no-locks --consistency lock`
+	err = ParseExtraArgs(&logger, exportCfg, strings.Fields(extraArgs))
+	c.Assert(err.Error(), Equals, "cannot both specify `--no-locks` and `--consistency` other than `none`")
 }
